requirements.txt
streamlit
pandas
numpy
yfinance
crewai
crewai_tools
langchain
langchain-huggingface
python-dotenv